Project/Project1-ChatBot/requirements.txt
langchain
langgraph==0.3.1 
langchain_community
langchain_core
langchain_groq
faiss-cpu
streamlit
tavily-python